[Fixed] Syntax Warning

+++ drivers/usb/dwc3/debug_ipc.c
@@ -164,7 +164,7 @@ void dwc3_dbg_dma_map(struct dwc3 *dwc, u8 ep_num, struct dwc3_request *req)
 
 	ipc_log_string(dwc->dwc_dma_ipc_log_ctxt,
 		"%02X-%-3.3s %-25.25s 0x%pK 0x%lx %u 0x%lx", ep_num >> 1,
-		ep_num & 1 ? "IN":"OUT", "MAP", &req->request, req->request.dma,
+		ep_num & 1 ? "IN":"OUT", "MAP", &req->request, (long int)req->request.dma,
 		req->request.length, (long int)req->trb_dma);
 }
 
@@ -176,7 +176,7 @@ void dwc3_dbg_dma_dequeue(struct dwc3 *dwc, u8 ep_num, struct dwc3_request *req)
 	ipc_log_string(dwc->dwc_dma_ipc_log_ctxt,
 		"%02X-%-3.3s %-25.25s 0x%pK 0x%lx 0x%lx", ep_num >> 1,
 		ep_num & 1 ? "IN":"OUT", "DEQUEUE", &req->request,
-		req->request.dma, req->trb_dma);
+		(long int)req->request.dma, (long int)req->trb_dma);
 }
 
 void dwc3_dbg_dma_queue(struct dwc3 *dwc, u8 ep_num, struct dwc3_request *req)
